Blocking out a TS impl of the Pagefind JS bindings
