+++ pagefind_web_js/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@pagefind/js",
+  "version": "0.0.0",
+  "type": "module",
+  "main": "npm_dist/cjs/js-api.cjs",
+  "module": "npm_dist/mjs/js-api.mjs",
+  "scripts": {
+    "build-coupled": "node ./build.js"
+  },
+  "author": "Cloudcannon",
+  "license": "MIT",
+  "devDependencies": {
+    "esbuild": "^0.18.2",
+    "typescript": "^5.1.3"
+  },
+  "files": [
+    "src/types/**"
+  ]
+}
