+++ pagefind_web_js/tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+        "allowJs": true,
+        "checkJs": true,
+        "noEmit": true,
+        "strict": true,
+        "target": "es2020",
+        "module": "es2022",
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true,
+        "paths": {
+            "pagefindWeb": [
+                "./types/index"
+            ],
+            "pagefindWebInternal": [
+                "./types/internal"
+            ],
+        },
+        "noUnusedLocals": true,
+        "noUnusedParameters": true
+    },
+    "include": [
+        "lib/**/*"
+    ]
+}
