+++ pagefind_web_js/types/internal.d.ts
@@ -1,0 +1,12 @@
+import "pagefindWeb"
+
+export type PagefindEntryJson = {
+    version: string,
+    languages: Record<string, PagefindEntryLanguage>,
+}
+
+export type PagefindEntryLanguage = {
+    hash: string,
+    wasm?: string,
+    page_count: number,
+}
